Move to ext. Fix build
