+++ CMakeLists.txt
@@ -250,7 +250,4 @@ setup_install_headers(${FLASHLIGHT_EXT_DIR} ${FL_INSTALL_INC_DIR_HEADER_LOC})
 set(FLASHLIGHT_APPS_DIR "${PROJECT_SOURCE_DIR}/app")
 include(${FLASHLIGHT_APPS_DIR}/CMakeLists.txt)
 
-set(FLASHLIGHT_TOOLS_DIR "${PROJECT_SOURCE_DIR}/tools")
-include(${FLASHLIGHT_TOOLS_DIR}/CMakeLists.txt)
-
 setup_install_targets(INSTALL_TARGETS ${INSTALLABLE_TARGETS})
