+++ ext/image/CMakeLists.txt
@@ -1,0 +1,5 @@
+cmake_minimum_required(VERSION 3.5.1)
+
+include(${CMAKE_CURRENT_LIST_DIR}/af/CMakeLists.txt)
+include(${CMAKE_CURRENT_LIST_DIR}/fl/CMakeLists.txt)
+
