+++ ext/image/af/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.5.1)
 
 target_sources(
-  flashlight-tools-image-af
+  flashlight
   PRIVATE
   ${CMAKE_CURRENT_LIST_DIR}/Transforms.cpp
   ${CMAKE_CURRENT_LIST_DIR}/Jpeg.cpp
