+++ ext/image/fl/models/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.5.1)
 
 target_sources(
-  flashlight-tools-image-fl
+  flashlight
   PRIVATE
   ${CMAKE_CURRENT_LIST_DIR}/Resnet.cpp
 )
