+++ ext/image/fl/nn/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.5.1)
 
 target_sources(
-  flashlight-tools-image-fl
+  flashlight
   PRIVATE
   # TODO (padentomasello) seperate src / header
   ${CMAKE_CURRENT_LIST_DIR}/Transformer.h
