+++ ext/image/fl/nn/PositionalEmbeddingSine.h
